Attempt consistency in root filesystem label
There's something weird going on with the root filesystem label.
ExtDiskMount prefixes fslabel with "_".  It's not clear why, but in
trying to solve some problems booting the AMI built with this image I
was frustrated by the inconsistency and wanted to make it a little
more transparent (although I didn't make it any less dumb).